+++ src/core/client/startup.ts
@@ -45,12 +45,15 @@ import './systems/tick';
 import './systems/toolbar';
 import './systems/vehicle';
 import './systems/world';
+
 // Utility
 import './utility/ipl';
 import './utility/notification';
 import './utility/reload';
 import './utility/screenshot';
+import './utility/spinner';
 import './utility/wheelMenu';
+
 // Views
 import './views/atm';
 import './views/characters';